# dist/swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Nutech PPOB Server",
        "description": "REST API server that handles users, transactions, and information related feature",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://nutech-ppob-server.vercel.app/"
        },
        {
            "url": "http://localhost:3000"
        }
    ],
...
    "paths": {
...
        "/v1/profile": {
            "get": {
                "tags": [
                    "Module Membership"
                ],
                "summary": "Get profile for current user",
                "operationId": "getProfileForCurrentUser",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/profile/update": {
            "put": {
                "tags": [
                    "Module Membership"
                ],
                "summary": "Update profile for current user",
                "operationId": "updateProfileForCurrentUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/userUpdate"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "userUpdate": {
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string",
                        "example": "John Edited"
                    },
                    "lastName": {
                        "type": "string",
                        "example": "Doe Edited"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}